* Preferences: Reset port to previous value if failed (not integer, start greater than end)  * PortScan: Port as integer instead of long (unisgned, 1-65535)  * PortScan: timeout is handled by ports group, and no more for every port (performance)  * PortScan: timeout is obtained by ScanRate.getRate(), host based.  * PortScan: select timeout is 100ms instead of 500, socket timeout is 1000 instead of 1500 (by default)  * ScanRate: responseTime is long instead of float  * ScanRate: Add 30% of the responseTime which defines the rate

+++ src/info/lamatricexiste/network/Utils/Export.java
@@ -70,9 +70,9 @@ public class Export {
                         + host.getHardwareAddress() + "\">\r\n";
                 // Open Ports //TODO: rething the XML structure to include close
                 // and filtered ports
-                long[] ports = host.getPortsOpen();
+                int[] ports = host.getPortsOpen();
                 if (ports != null) {
-                    for (long port : ports) {
+                    for (int port : ports) {
                         xml += "\t\t\t<port>" + port + "/tcp open</port>\r\n";
                     }
                 }
